Prereq: move EventRegistry to subnamespace
As new registries will be created (for command handlers), splitting them in
sub namespaces is valueable
