+++ tests/Registry/Event/Cached/CachedEventRegistryDecoratorTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Gember\EventSourcing\Test\Registry\Cached;
+namespace Gember\EventSourcing\Test\Registry\Event\Cached;
 
-use Gember\EventSourcing\Registry\Cached\CachedEventRegistryDecorator;
-use Gember\EventSourcing\Registry\Reflector\ReflectorEventRegistry;
+use Gember\EventSourcing\Registry\Event\Cached\CachedEventRegistryDecorator;
+use Gember\EventSourcing\Registry\Event\Reflector\ReflectorEventRegistry;
 use Gember\EventSourcing\Resolver\DomainEvent\NormalizedEventName\Attribute\AttributeNormalizedEventNameResolver;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseModifiedEvent;
